fix : 화면이 이상하게 보이는 버그 해결
