+++ src/main/java/wooteco/chess/domain/ChessGame.java
@@ -5,39 +5,39 @@ import wooteco.chess.domain.state.ChessGameState;
 import wooteco.chess.domain.state.Score;
 
 public class ChessGame {
-	private ChessGameState state;
+    private ChessGameState state;
 
-	public ChessGame(ChessGameState state) {
-		this.state = state;
-	}
+    public ChessGame(ChessGameState state) {
+        this.state = state;
+    }
 
-	public void start() {
-		state = state.start();
-	}
+    public void start() {
+        state = state.start();
+    }
 
-	public void move(Position source, Position target) {
-		state = state.move(source, target);
-	}
+    public void move(Position source, Position target) {
+        state = state.move(source, target);
+    }
 
-	public void end() {
-		state = state.end();
-	}
+    public void end() {
+        state = state.end();
+    }
 
-	public Board board() {
-		return state.board();
-	}
+    public Board board() {
+        return state.board();
+    }
 
-	public boolean isEnd() {
-		return state.isEnd();
-	}
+    public boolean isEnd() {
+        return state.isEnd();
+    }
 
-	public Turn turn() {
-		return state.turn();
-	}
+    public Turn turn() {
+        return state.turn();
+    }
 
-	public Status status() {
-		Score white = state.score(Team.WHITE);
-		Score black = state.score(Team.BLACK);
-		return new Status(white, black);
-	}
+    public Status status() {
+        Score white = state.score(Team.WHITE);
+        Score black = state.score(Team.BLACK);
+        return new Status(white, black);
+    }
 }
