+++ src/main/resources/static/js/chess.js
@@ -1,5 +1,6 @@
 function on_games() {
-    fetch("http://localhost:4567/games").then(response => response.json()).then(playersDto => {
-        console.log(playersDto);
-    })
+  fetch("http://localhost:4567/room/games").then(response => response.json()).then(playersDto => {
+    console.log(playersDto.player1Name);
+    console.log(playersDto);
+  })
 }
